refactor: 70.J.3 — split freminal/src/gui/mod.rs into sibling modules
mod.rs reduced from 3,244 to 410 lines by extracting six sibling
files under freminal/src/gui/ (same pattern as Task 70.J.1/70.J.2):

- tab_spawning.rs (375 lines) — spawn_new_tab, spawn_split_pane,
  spawn_new_window, spawn_pane_from_leaf, initial_size_for_split
- layout_ops.rs (604 lines) — save_layout, apply_layout, build_tab_*,
  build_window_*, build_tabs_for_window, insert_subtree_as_split,
  inject_layout_commands, read_cwd_for_pane_with_extra, plus the
  layout_dir_to_pane_dir and extract_root_leaf free helpers
- settings_dispatch.rs (197 lines) — handle_settings_action
- session.rs (141 lines) — last_session_path, persist_window_state,
  snapshot_main_window_geometry, auto_save_session,
  maybe_restore_last_session
- app_impl.rs (1,547 lines) — full impl freminal_windowing::App for
  FreminalGui (on_window_created, on_close_requested, clear_color,
  update, raw_input_hook)
- run.rs (81 lines) — pub fn run entry point, re-exported via
  `pub use run::run;` so crate::gui::run resolves unchanged

Pure move refactor — no signatures, logic, or visibility changed.
Two free helpers promoted to pub(super) so sibling modules can
import them. Test module remains in mod.rs.

Verification: cargo test --all (all 4913+ tests pass), cargo clippy
--all-targets --all-features -- -D warnings (clean), cargo machete
(clean).
